app/requirements.txt
fastapi
uvicorn
numpy
Pillow
torchvision
tritonclient[http]
python-multipart